fix: fix id and abstract class

+++ SBA_M/src/main/java/com/example/SBA_M/controller/RoleController.java
@@ -22,7 +22,7 @@ public class RoleController {
     }
 
     @GetMapping("/{id}")
-    public Role getRoleById(@PathVariable Long id) {
+    public Role getRoleById(@PathVariable Integer id) {
         return roleService.getRoleById(id);
     }
 
@@ -32,12 +32,12 @@ public class RoleController {
     }
 
     @PutMapping("/{id}")
-    public Role updateRole(@PathVariable Long id, @RequestBody Role role) {
+    public Role updateRole(@PathVariable Integer id, @RequestBody Role role) {
         return roleService.updateRole(id, role);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteRole(@PathVariable Long id) {
+    public void deleteRole(@PathVariable Integer id) {
         roleService.deleteRole(id);
     }
 }
