+++ SBA_M/src/main/java/com/example/SBA_M/entity/commands/UniversityCategory.java
@@ -1,6 +1,5 @@
 package com.example.SBA_M.entity.commands;
 
-import com.example.SBA_M.entity.AbstractEntity;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
@@ -13,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
-public class UniversityCategory extends AbstractEntity<Long> {
+public class UniversityCategory extends AbstractEntity<Integer> {
 
     @Column(nullable = false, unique = true, length = 100)
     private String name;
