+++ SBA_M/src/main/java/com/example/SBA_M/repository/commands/UniversityCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversityCategoryRepository extends JpaRepository<UniversityCategory, Long> {
+public interface UniversityCategoryRepository extends JpaRepository<UniversityCategory, Integer> {
 }
