+++ SBA_M/src/main/java/com/example/SBA_M/repository/queries/UniversityReadRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversityReadRepository extends MongoRepository<UniversityDocument, Long> {
+public interface UniversityReadRepository extends MongoRepository<UniversityDocument, Integer> {
 }
